Fix typo in Annotation - color control
In the Annotation section, the 'color' description says you can add color presets. This is listed as a `presetsColors` in the controls, but I tried to use this in my project and it doesn't work. I think this is a typo for `presetColors` (this spelling does work for me locally).